Fixes, cleanup, move generateChildren out of demo.js
