--- demo/initialMusicData.js
@@ -1,81 +1,0 @@
-export default [
-  [
-     {
-       score: 0,
-       gain: 0,
-       sample: "samples/kick.wav",
-       beat: [1, 0, 0, 0, 0, 0, 0, 0],
-     },
-     {
-      score: 0,
-      gain: 0,
-       sample: "samples/8580__ST_060.wav",
-       beat: [0, 0, 0, 0, 1, 0, 0, 0],
-     },
-     {
-      score: 0,
-      gain: 0,
-       sample: "samples/hihat.wav",
-       beat: [1, 1, 1, 1, 1, 1, 1, 1], 
-     },
-  ],
-  [
-     {
-       score: 0,
-       gain: 0,
-       sample: "samples/kick.wav",
-       beat: [1, 1, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0],
-     },
-     {
-      score: 0,
-      gain: 0,
-       sample: "samples/8580_PST_090.wav",
-       beat: [0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0],
-     },
-
-  ],
-
-  [
-     {
-       score: 0,
-       gain: 0,
-       sample: "samples/kick.wav",
-       beat: [1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0],
-     },
-     {
-      score: 0,
-      gain: 0,
-       sample: "samples/cowbell.wav",
-       beat: [0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0],
-     },
-     {
-      score: 0,
-      gain: 0,
-       sample: "samples/hihat.wav",
-       beat: [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-     },
-
-  ],
-  [
-     {
-       score: 0,
-       gain: 0,
-       sample: "samples/kick.wav",
-       beat: [1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0],
-     },
-     {
-      score: 0,
-      gain: 0,
-       sample: "samples/cowbell.wav",
-       beat: [1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1],
-     },
-     {
-      score: 0,
-      gain: 0,
-       sample: "samples/hihat.wav",
-       beat: [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-     },
-
-  ] 
-
-]
